Initial commit: BPF-Allocator-Bench
BPF-Allocator-Bench is a bare mininum project to stress and benchmark
the eBPF JIT memory allocator. It includes a bpf program that has 8 BPF
programs. 6 are tracepoints and 2 are kprobes. The userspace program
loads these programs and calls 5 syscalls in a loop. This triggers 5/8
BPF programs continously.

The wrapper run.sh script creates 20 instances of this stress program
which loads 8 * 20 = 160 BPF programs and then triggers 100 of them
continously. It then runs perf to measure iTLB related metrics.

The main usecase of this benchmark is to understand how different ways
of allocation memory for BPF programs can improve iTLB performance.

Signed-off-by: Puranjay Mohan <puranjay12@gmail.com>